+++ routes/UserRoutes.js
@@ -1,5 +1,5 @@
 import Express  from "express";
-import { getAllUsers, register } from "../controllers/User.js";
+import { getAllUsers, register, login } from "../controllers/User.js";
 
 const Router = Express.Router();
 
@@ -7,6 +7,6 @@ Router.get('/', getAllUsers);
 
 Router.post('/new', register);
 
-// Router.post('/old', login);
+Router.post('/old', login);
 
 export default Router;